*cries*

+++ src/docker.js
@@ -3,11 +3,14 @@ const { execSync } = require('child_process');
 const path = require('path');
 const fs = require('fs');
 
-var Dockerfile = `FROM node:18
+const Dockerfile = `FROM node:20.5.1-bookworm-slim
+RUN groupadd --gid 420 {username} && useradd --uid 10420 --gid {username} --shell /bin/bash --create-home {username}
+RUN apt-get update && apt-get install -y --no-install-recommends dumb-init
 WORKDIR /home/{username}/{name}
-COPY package*.json ./
-RUN npm install
-COPY . .
+COPY --chown={username}:{username} package*.json ./
+RUN {install}
+COPY --chown={username}:{username} . .
+USER {username}
 CMD {cmd}`
 
 /** @returns {Promise<string[]>} */
@@ -16,18 +19,19 @@ async function getLogs(id) {
   return logs;
 }
 
-async function createDocker(file, username, env) {
+async function createDocker(file, username, envArr = []) {
   if(file.includes('.')) file = file.split('.')[0];
   const cmd = [];
+  /** @type {string} */
   var docker = Dockerfile
-    .replace('{username}', username)
+    .replaceAll(/{username}/g, username)
     .replace('{name}', file)
 
-  if(env.length > 0) env.forEach(envVar => {
-    cmd.push(`${envVar.name}=${envVar.value}`);
+  if(envArr.length > 0) envArr.forEach(env => {
+    cmd.push(`${env.name}=${env.value}`);
   })
 
-  cmd.push('node', '.');
+  cmd.push('dumb-init', 'node', '.');
   docker = docker.replace(/{cmd}/, JSON.stringify(cmd).replace(/'/gm, ''));
 
 
@@ -36,7 +40,9 @@ async function createDocker(file, username, env) {
   if(!fs.existsSync(path.join(__dirname, 'temp', username))) await fs.mkdirSync(path.join(__dirname, 'temp', username), { recursive: true });
 
   return new Promise(async (resolve, reject) => {
-    await execSync(`unzip ./src/verify/${username}/${file} -d ${pth}`);
+    await execSync(`unzip -o ./src/verify/${username}/${file} -d ${pth}`);
+    if(await fs.existsSync(`${pth}/package-lock.json`)) docker = docker.replace('{install}', 'npm ci --only=production');
+    else docker = docker.replace('{install}', "npm install --omit=dev");
     await fs.writeFileSync(`${pth}/Dockerfile`, docker);
     await execSync(`cp .dockerignore ${pth}`, { cwd: dckr });
     await execSync(`docker build . -t ${username}/${file}`, { cwd: pth });
@@ -58,7 +64,7 @@ async function startDocker({username, name, id}) {
   if(id) {
     execSync(`docker start ${id}`)
   } else {
-    const fullId = execSync(`docker run -d --memory="512m" --cpus="1" ${username}/${name}`).toString().trim();
+    const fullId = execSync(`docker run -d --memory="512m" --cpus="0.25" ${username}/${name}`).toString().trim();
     const id = fullId.slice(0, 11);
 
     return id;
@@ -79,13 +85,13 @@ async function close( username, name ) {
   return user.dockers[name].id;
 }
 
-async function deleteDocker( username, name ) {
+async function deleteDocker( username, name, id ) {
   try {
-    var id = await close(username, name);
+    await close(username, name);
   } catch (e) {}
 
   execSync(`docker rmi -f ${username}/${name}`);
-  execSync(`docker rm -f ${id}`);
+  if(typeof id == 'string' && id !== "") execSync(`docker rm -f ${id}`);
 }
 
 module.exports = {
